downtrack: send silence on audio mute to avoid comfort noise static
When DTX is enabled, muting an audio track can cause a static hiss as
the receiver attempts to fill the missing audio packets with comfort
noise. When muting, forward some empty silence packets first so DTX does
not attempt to fill in the audio with the last sent sound.

Fixes #675